+++ service/src/main/java/org/kofi/creditex/service/CommitteeService.java
@@ -6,9 +6,9 @@ import java.util.ArrayList;
 import java.sql.Date;
 
 public interface CommitteeService {
-    Application GetApplication(int application_id);
-    List<Vote> GetApplicationVotes(int application_id);
+    Application GetApplication(long application_id);
+    List<Vote> GetApplicationVotes(long application_id);
     List<Application> GetVotingApplications(boolean voting);
-    int Vote(String committee_name, int application_id, boolean acceptance, String comment);
-    List<Credit> GetClientCredits(int client_id);
+    int Vote(String committee_name, long application_id, boolean acceptance, String comment);
+    List<Credit> GetClientCredits(long client_id);
 }
