+++ service/src/main/java/org/kofi/creditex/service/CreditService.java
@@ -4,6 +4,6 @@ import org.kofi.creditex.model.Credit;
 import org.kofi.creditex.web.model.CreditForm;
 
 public interface CreditService {
-    Credit GetCreditById(int id);
-    CreditForm GetCreditFormById(int id);
+    Credit GetCreditById(long id);
+    CreditForm GetCreditFormById(long id);
 }
