+++ service/src/main/java/org/kofi/creditex/service/SecurityService.java
@@ -9,11 +9,11 @@ public interface SecurityService {
     List<Application> GetSecurityApplications();
     List<Credit> GetExpiredCredits();
     List<Credit> GetUnreturnedCredits();
-    boolean ConfirmApplication(String security_name, int application_id, boolean confirmation, String comment);
-    Application GetApplication(int id);
-    List<Credit> GetCurrentClientCredits(int client_id);
-    List<Credit> GetClientExpiredCredits(int client_id);
-    List<Credit> GetClientUnreturnedCredits(int client_id);
-    List<PriorRepaymentApplication> GetClientPriorRepaymentApplications(int client_id);
-    List<ProlongationApplication> GetClientProlongationApplications(int client_id);
+    boolean ConfirmApplication(String security_name, long application_id, boolean confirmation, String comment);
+    Application GetApplication(long id);
+    List<Credit> GetCurrentClientCredits(long client_id);
+    List<Credit> GetClientExpiredCredits(long client_id);
+    List<Credit> GetClientUnreturnedCredits(long client_id);
+    List<PriorRepaymentApplication> GetClientPriorRepaymentApplications(long client_id);
+    List<ProlongationApplication> GetClientProlongationApplications(long client_id);
 }
